New module specphot_kernels_merging.py
