+++ diffsky/experimental/kernels/mc_phot_kernels.py
@@ -17,6 +17,7 @@ from . import (
     phot_kernels,
     phot_kernels_merging,
     sed_kernels,
+    specphot_kernels_merging,
 )
 from . import ssp_weight_kernels as sspwk
 
@@ -39,7 +40,8 @@ _sed_kern = sed_kernels._sed_kern
 _get_dbk_linelum_decomposition = dbk_kernels._get_dbk_linelum_decomposition
 _mc_phot_kern_merging = phot_kernels_merging._mc_phot_kern_merging
 _phot_kern_merging = phot_kernels_merging._phot_kern_merging
-
+_mc_specphot_kern_merging = specphot_kernels_merging._mc_specphot_kern_merging
+_specphot_kern_merging = specphot_kernels_merging._specphot_kern_merging
 
 # randoms
 get_mc_phot_randoms = mc_randoms.get_mc_phot_randoms
@@ -51,161 +53,6 @@ SpecKernResults = linelum_kernels.SpecKernResults
 PhotKernResults = phot_kernels.PhotKernResults
 DBKRandoms = mc_randoms.DBKRandoms
 SEDKernResults = sed_kernels.SEDKernResults
-
-
-@jjit
-def _mc_specphot_kern_merging(
-    ran_key,
-    phot_randoms,
-    sfh_params,
-    z_obs,
-    t_obs,
-    mah_params,
-    ssp_data,
-    precomputed_ssp_mag_table,
-    z_phot_table,
-    wave_eff_table,
-    line_wave_table,
-    diffstarpop_params,
-    mzr_params,
-    spspop_params,
-    scatter_params,
-    ssp_err_pop_params,
-    merge_params,
-    cosmo_params,
-    fb,
-    logmp_infall,
-    logmhost_infall,
-    t_infall,
-    is_central,
-    nhalos_weights,
-    halo_indx,
-):
-    phot_randoms, sfh_params = get_mc_phot_randoms(
-        ran_key, diffstarpop_params, mah_params, cosmo_params
-    )
-
-    (
-        phot_kern_results,
-        linelums_in_situ,
-        flux_obs,
-        merge_prob,
-        mstar_obs,
-        linelums_obs,
-    ) = _specphot_kern_merging(
-        phot_randoms,
-        sfh_params,
-        z_obs,
-        t_obs,
-        mah_params,
-        ssp_data,
-        precomputed_ssp_mag_table,
-        z_phot_table,
-        wave_eff_table,
-        line_wave_table,
-        mzr_params,
-        spspop_params,
-        scatter_params,
-        ssp_err_pop_params,
-        merge_params,
-        cosmo_params,
-        fb,
-        logmp_infall,
-        logmhost_infall,
-        t_infall,
-        is_central,
-        nhalos_weights,
-        halo_indx,
-    )
-    return (
-        phot_kern_results,
-        linelums_in_situ,
-        phot_randoms,
-        flux_obs,
-        merge_prob,
-        mstar_obs,
-        linelums_obs,
-    )
-
-
-@jjit
-def _specphot_kern_merging(
-    phot_randoms,
-    sfh_params,
-    z_obs,
-    t_obs,
-    mah_params,
-    ssp_data,
-    precomputed_ssp_mag_table,
-    z_phot_table,
-    wave_eff_table,
-    line_wave_table,
-    mzr_params,
-    spspop_params,
-    scatter_params,
-    ssp_err_pop_params,
-    merge_params,
-    cosmo_params,
-    fb,
-    logmp_infall,
-    logmhost_infall,
-    t_infall,
-    is_central,
-    nhalos_weights,
-    halo_indx,
-):
-    phot_kern_results, linelums_in_situ, dust_ftrans_lines = _specphot_kern(
-        phot_randoms,
-        sfh_params,
-        z_obs,
-        t_obs,
-        mah_params,
-        ssp_data,
-        precomputed_ssp_mag_table,
-        z_phot_table,
-        wave_eff_table,
-        line_wave_table,
-        mzr_params,
-        spspop_params,
-        scatter_params,
-        ssp_err_pop_params,
-        cosmo_params,
-        fb,
-    )
-
-    upids = jnp.where(is_central == 1, -1.0, 0.0)
-    merge_prob = merging_model.get_p_merge_from_merging_params(
-        merge_params, logmp_infall, logmhost_infall, t_obs, t_infall, upids
-    )
-
-    mstar_in_situ = 10**phot_kern_results.logsm_obs
-    mstar_obs = compute_x_tot_from_x_in_situ(
-        mstar_in_situ, merge_prob, nhalos_weights, halo_indx
-    )
-
-    flux_in_situ = 10 ** (-0.4 * phot_kern_results.obs_mags)
-    flux_obs = compute_x_tot_from_x_in_situ(
-        flux_in_situ,
-        merge_prob[:, jnp.newaxis],
-        nhalos_weights[:, jnp.newaxis],
-        halo_indx,
-    )
-
-    linelums_obs = compute_x_tot_from_x_in_situ(
-        linelums_in_situ,
-        merge_prob[:, jnp.newaxis],
-        nhalos_weights[:, jnp.newaxis],
-        halo_indx,
-    )
-
-    return (
-        phot_kern_results,
-        linelums_in_situ,
-        flux_obs,
-        merge_prob,
-        mstar_obs,
-        linelums_obs,
-    )
 
 
 def _mc_dbk_phot_kern(
