+++ diffsky/experimental/kernels/specphot_kernels_merging.py
@@ -1,0 +1,163 @@
+""""""
+
+from jax import jit as jjit
+from jax import numpy as jnp
+
+from ...merging import compute_x_tot_from_x_in_situ, merging_model
+from . import linelum_kernels, mc_randoms
+
+
+@jjit
+def _mc_specphot_kern_merging(
+    ran_key,
+    phot_randoms,
+    sfh_params,
+    z_obs,
+    t_obs,
+    mah_params,
+    ssp_data,
+    precomputed_ssp_mag_table,
+    z_phot_table,
+    wave_eff_table,
+    line_wave_table,
+    diffstarpop_params,
+    mzr_params,
+    spspop_params,
+    scatter_params,
+    ssp_err_pop_params,
+    merge_params,
+    cosmo_params,
+    fb,
+    logmp_infall,
+    logmhost_infall,
+    t_infall,
+    is_central,
+    nhalos_weights,
+    halo_indx,
+):
+    phot_randoms, sfh_params = mc_randoms.get_mc_phot_randoms(
+        ran_key, diffstarpop_params, mah_params, cosmo_params
+    )
+
+    (
+        phot_kern_results,
+        linelums_in_situ,
+        flux_obs,
+        merge_prob,
+        mstar_obs,
+        linelums_obs,
+    ) = _specphot_kern_merging(
+        phot_randoms,
+        sfh_params,
+        z_obs,
+        t_obs,
+        mah_params,
+        ssp_data,
+        precomputed_ssp_mag_table,
+        z_phot_table,
+        wave_eff_table,
+        line_wave_table,
+        mzr_params,
+        spspop_params,
+        scatter_params,
+        ssp_err_pop_params,
+        merge_params,
+        cosmo_params,
+        fb,
+        logmp_infall,
+        logmhost_infall,
+        t_infall,
+        is_central,
+        nhalos_weights,
+        halo_indx,
+    )
+    return (
+        phot_kern_results,
+        linelums_in_situ,
+        phot_randoms,
+        flux_obs,
+        merge_prob,
+        mstar_obs,
+        linelums_obs,
+    )
+
+
+@jjit
+def _specphot_kern_merging(
+    phot_randoms,
+    sfh_params,
+    z_obs,
+    t_obs,
+    mah_params,
+    ssp_data,
+    precomputed_ssp_mag_table,
+    z_phot_table,
+    wave_eff_table,
+    line_wave_table,
+    mzr_params,
+    spspop_params,
+    scatter_params,
+    ssp_err_pop_params,
+    merge_params,
+    cosmo_params,
+    fb,
+    logmp_infall,
+    logmhost_infall,
+    t_infall,
+    is_central,
+    nhalos_weights,
+    halo_indx,
+):
+    _res = linelum_kernels._specphot_kern(
+        phot_randoms,
+        sfh_params,
+        z_obs,
+        t_obs,
+        mah_params,
+        ssp_data,
+        precomputed_ssp_mag_table,
+        z_phot_table,
+        wave_eff_table,
+        line_wave_table,
+        mzr_params,
+        spspop_params,
+        scatter_params,
+        ssp_err_pop_params,
+        cosmo_params,
+        fb,
+    )
+    phot_kern_results, linelums_in_situ, dust_ftrans_lines = _res
+
+    upids = jnp.where(is_central == 1, -1.0, 0.0)
+    merge_prob = merging_model.get_p_merge_from_merging_params(
+        merge_params, logmp_infall, logmhost_infall, t_obs, t_infall, upids
+    )
+
+    mstar_in_situ = 10**phot_kern_results.logsm_obs
+    mstar_obs = compute_x_tot_from_x_in_situ(
+        mstar_in_situ, merge_prob, nhalos_weights, halo_indx
+    )
+
+    flux_in_situ = 10 ** (-0.4 * phot_kern_results.obs_mags)
+    flux_obs = compute_x_tot_from_x_in_situ(
+        flux_in_situ,
+        merge_prob[:, jnp.newaxis],
+        nhalos_weights[:, jnp.newaxis],
+        halo_indx,
+    )
+
+    linelums_obs = compute_x_tot_from_x_in_situ(
+        linelums_in_situ,
+        merge_prob[:, jnp.newaxis],
+        nhalos_weights[:, jnp.newaxis],
+        halo_indx,
+    )
+
+    return (
+        phot_kern_results,
+        linelums_in_situ,
+        flux_obs,
+        merge_prob,
+        mstar_obs,
+        linelums_obs,
+    )
